Change the data type from int to long in verifyInt method in Item class Add the check 13 length digit for isbn checking in book class. Modify the README.md file.

+++ src/LibraryManagment/Item.java
@@ -46,7 +46,7 @@ public abstract class Item {
     // Method to verify the string in the correct format.
     public static boolean verifyInt(String integer) {
         try {
-            int a = Integer.parseInt(integer);
+            long a = Long.parseLong(integer); // Since the 13 digit is out of integer's range, it is compulsory to use long data type.
             return true;
         } catch (NumberFormatException e) {
             return false;
